requirements.txt
click>=8.1.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.5.0
ollama>=0.1.0
rich>=13.7.0
email-validator>=1.3.0
